requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
pydantic==2.9.2
python-multipart==0.0.9
python-docx==1.1.2
PyMuPDF==1.24.10
openai==1.51.0
python-dotenv==1.0.1
pytest==8.3.3
httpx==0.27.2
